update giao diện + hoàn thiện product

+++ WEBSITE_BAN_QUAN_AO_NAM/src/main/java/com/poly/datn/service/impl/ImageServiceImpl.java
@@ -12,6 +12,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 @Service
@@ -25,14 +27,20 @@ public class ImageServiceImpl implements ImageService {
 
 
     @Override
-    public void deleteImage(Long id) {
-        Image image = repository.getReferenceById(id);
+    public Boolean deleteImage(Long id) {
+        Image image = repository.findById(id).orElse(null);
 
-        try {
-            cloudinary.uploader().destroy(image.getPublicId(), ObjectUtils.emptyMap());
+        if (image != null) {
+            try {
+                cloudinary.uploader().destroy(image.getPublicId(), ObjectUtils.emptyMap());
+                repository.delete(image);
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
             repository.delete(image);
-        } catch (IOException e) {
-            e.printStackTrace();
+            return true; // Trả về true để chỉ ra xóa thành công.
+        } else {
+            return false; // Trả về false để chỉ ra không tìm thấy ảnh với ID tương ứng.
         }
     }
 
@@ -57,4 +65,19 @@ public class ImageServiceImpl implements ImageService {
             e.printStackTrace();
         }
     }
+
+    @Override
+    public Map<?, ?> upload(MultipartFile file, String foder) throws IOException {
+        Map<?, ?> uploadResult = new HashMap<>();
+        if (file.isEmpty()) return uploadResult;
+        Map<String, Object> params = ObjectUtils.asMap("folder", foder);
+         uploadResult = cloudinary.uploader().upload(file.getBytes(), params);
+        return uploadResult;
+    }
+
+    @Override
+    public List<Image> getListanh(Long id) {
+        return repository.getListAnh(id);
+    }
+
 }
